requirements.txt
selenium
pymongo
flask
python-dotenv
webdriver_manager
requests 